Merge 1.0.8

+++ app/src/main/java/eu/kanade/tachiyomi/data/download/DownloadManager.kt
@@ -10,7 +10,11 @@ import eu.kanade.tachiyomi.data.download.model.DownloadQueue
 import eu.kanade.tachiyomi.source.Source
 import eu.kanade.tachiyomi.source.SourceManager
 import eu.kanade.tachiyomi.source.model.Page
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import rx.Observable
+import timber.log.Timber
 import uy.kohesive.injekt.injectLazy
 
 /**
@@ -227,29 +231,42 @@ class DownloadManager(val context: Context) {
      * @param source the source of the chapters.
      */
     fun deleteChapters(chapters: List<Chapter>, manga: Manga, source: Source) {
-        val wasPaused = isPaused()
-        if (chapters.isEmpty()) {
-            DownloadService.stop(context)
-            downloader.queue.clear()
-            return
+        GlobalScope.launch(Dispatchers.IO) {
+            val wasPaused = isPaused()
+            if (chapters.isEmpty()) {
+                DownloadService.stop(context)
+                downloader.queue.clear()
+                return@launch
+            }
+            downloader.pause()
+            downloader.queue.remove(chapters)
+            if (!wasPaused && downloader.queue.isNotEmpty()) {
+                downloader.start()
+            } else if (downloader.queue.isEmpty() && DownloadService.isRunning(context)) {
+                DownloadService.stop(context)
+            } else if (downloader.queue.isEmpty()) {
+                DownloadService.callListeners(false)
+                downloader.stop()
+            }
+            queue.remove(chapters)
+            val chapterDirs =
+                provider.findChapterDirs(chapters, manga, source) + provider.findTempChapterDirs(
+                    chapters, manga, source
+                )
+            chapterDirs.forEach { it.delete() }
+            cache.removeChapters(chapters, manga)
+            if (cache.getDownloadCount(manga, true) == 0) { // Delete manga directory if empty
+                chapterDirs.firstOrNull()?.parentFile?.delete()
+            }
+            queue.updateListeners()
         }
-        downloader.pause()
-        downloader.queue.remove(chapters)
-        if (!wasPaused && downloader.queue.isNotEmpty()) {
-            downloader.start()
-        } else if (downloader.queue.isEmpty() && DownloadService.isRunning(context)) {
-            DownloadService.stop(context)
-        } else if (downloader.queue.isEmpty()) {
-            DownloadService.callListeners(false)
-            downloader.stop()
-        }
-        queue.remove(chapters)
-        val chapterDirs = provider.findChapterDirs(chapters, manga, source) + provider.findTempChapterDirs(chapters, manga, source)
-        chapterDirs.forEach { it.delete() }
-        cache.removeChapters(chapters, manga)
-        if (cache.getDownloadCount(manga, true) == 0) { // Delete manga directory if empty
-            chapterDirs.firstOrNull()?.parentFile?.delete()
-        }
+    }
+
+    /**
+     * return the list of all manga folders
+     */
+    fun getMangaFolders(source: Source): List<UniFile> {
+        return provider.findSourceDir(source)?.listFiles()?.toList() ?: emptyList()
     }
 
     /**
@@ -259,19 +276,39 @@ class DownloadManager(val context: Context) {
      * @param manga the manga of the chapters.
      * @param source the source of the chapters.
      */
-    fun cleanupChapters(allChapters: List<Chapter>, manga: Manga, source: Source): Int {
+    fun cleanupChapters(allChapters: List<Chapter>, manga: Manga, source: Source, removeRead: Boolean, removeNonFavorite: Boolean): Int {
         var cleaned = 0
+
+        if (removeNonFavorite && !manga.favorite) {
+            val mangaFolder = provider.getMangaDir(manga, source)
+            cleaned += 1 + (mangaFolder.listFiles()?.size ?: 0)
+            mangaFolder.delete()
+            cache.removeManga(manga)
+            return cleaned
+        }
+
         val filesWithNoChapter = provider.findUnmatchedChapterDirs(allChapters, manga, source)
         cleaned += filesWithNoChapter.size
         cache.removeFolders(filesWithNoChapter.mapNotNull { it.name }, manga)
         filesWithNoChapter.forEach { it.delete() }
-        val readChapters = allChapters.filter { it.read }
-        val readChapterDirs = provider.findChapterDirs(readChapters, manga, source)
-        readChapterDirs.forEach { it.delete() }
-        cleaned += readChapterDirs.size
-        cache.removeChapters(readChapters, manga)
+
+        if (removeRead) {
+            val readChapters = allChapters.filter { it.read }
+            val readChapterDirs = provider.findChapterDirs(readChapters, manga, source)
+            readChapterDirs.forEach { it.delete() }
+            cleaned += readChapterDirs.size
+            cache.removeChapters(readChapters, manga)
+        }
+
         if (cache.getDownloadCount(manga) == 0) {
-            provider.findChapterDirs(allChapters, manga, source).firstOrNull()?.parentFile?.delete() // Delete manga directory if empty
+            val mangaFolder = provider.getMangaDir(manga, source)
+            val size = mangaFolder.listFiles()?.size ?: 0
+            if (size == 0) {
+                mangaFolder.delete()
+                cache.removeManga(manga)
+            } else {
+                Timber.e("Cache and download folder doesn't match for %s", manga.title)
+            }
         }
         return cleaned
     }
@@ -307,6 +344,27 @@ class DownloadManager(val context: Context) {
         for ((manga, chapters) in pendingChapters) {
             val source = sourceManager.get(manga.source) ?: continue
             deleteChapters(chapters, manga, source)
+        }
+    }
+
+    /**
+     * Renames an already downloaded chapter
+     *
+     * @param manga the manga of the chapter.
+     * @param oldChapter the existing chapter with the old name.
+     * @param newChapter the target chapter with the new name.
+     */
+    fun renameChapter(source: Source, manga: Manga, oldChapter: Chapter, newChapter: Chapter) {
+        val oldName = provider.getChapterDirName(oldChapter)
+        val newName = provider.getChapterDirName(newChapter)
+        val mangaDir = provider.getMangaDir(manga, source)
+
+        val oldFolder = mangaDir.findFile(oldName)
+        if (oldFolder?.renameTo(newName) == true) {
+            cache.removeChapters(listOf(oldChapter), manga)
+            cache.addChapter(newName, mangaDir, manga)
+        } else {
+            Timber.e("Could not rename downloaded chapter: %s.", oldName)
         }
     }
